requirements.txt
Django==5.0.2
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
Pillow==10.2.0
django-allauth==0.61.0
django-notifications-hq==1.8.3
python-dotenv==1.0.1
django-colorfield==0.11.0
django-widget-tweaks==1.5.0
whitenoise==6.6.0
django-compressor==4.4
django-browser-reload==1.12.1 